// features/api/model/auditLogResource.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { AuditLogResourceCauser } from './auditLogResourceCauser';
import type { AuditLogResourceChanges } from './auditLogResourceChanges';
import type { AuditLogResourceSubject } from './auditLogResourceSubject';

export interface AuditLogResource {
  id: string;
  description: string;
  time_ago: string;
  timestamp: string;
  causer: AuditLogResourceCauser;
  subject: AuditLogResourceSubject;
  changes: AuditLogResourceChanges;
}